services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
openai>=1.40.0
pydantic-settings>=2.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
